move check_for_idx_rename to do_alter_table
